+++ obj/src/persistence/IdentifiableMongoDbPersistence.js
@@ -36,9 +36,16 @@ const MongoDbPersistence_1 = require("./MongoDbPersistence");
  * - options:
  *   - max_pool_size:             (optional) maximum connection pool size (default: 2)
  *   - keep_alive:                (optional) enable connection keep alive (default: true)
- *   - connect_timeout:           (optional) connection timeout in milliseconds (default: 5 sec)
+ *   - connect_timeout:           (optional) connection timeout in milliseconds (default: 5000)
+ *   - socket_timeout:            (optional) socket timeout in milliseconds (default: 360000)
  *   - auto_reconnect:            (optional) enable auto reconnection (default: true)
+ *   - reconnect_interval:        (optional) reconnection interval in milliseconds (default: 1000)
  *   - max_page_size:             (optional) maximum page size (default: 100)
+ *   - replica_set:               (optional) name of replica set
+ *   - ssl:                       (optional) enable SSL connection (default: false)
+ *   - auth_source:               (optional) authentication source
+ *   - auth_user:                 (optional) authentication user name
+ *   - auth_password:             (optional) authentication user password
  *   - debug:                     (optional) enable debug output (default: false).
  *
  * ### References ###
@@ -101,16 +108,13 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
      * Creates a new instance of the persistence component.
      *
      * @param collection    (optional) a collection name.
-     * @param schema        (optional) a Mongoose schema.
      */
-    constructor(collection, schema) {
-        super(collection, schema);
+    constructor(collection) {
+        super(collection);
         //TODO (note for SS): is this needed? It's in MongoDbPersistence as well...
         this._maxPageSize = 100;
         if (collection == null)
             throw new Error("Collection name could not be null");
-        if (schema == null)
-            throw new Error("Schema could not be null");
     }
     /**
      * Configures component by passing configuration parameters.
@@ -149,25 +153,25 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
         let skip = paging.getSkip(-1);
         let take = paging.getTake(this._maxPageSize);
         let pagingEnabled = paging.total;
-        // Configure statement
-        let statement = this._model.find(filter);
+        // Configure options
+        let options = {};
         if (skip >= 0)
-            statement.skip(skip);
-        statement.limit(take);
+            options.skip = skip;
+        options.limit = take;
         if (sort && !_.isEmpty(sort))
-            statement.sort(sort);
+            options.sort = sort;
         if (select && !_.isEmpty(select))
-            statement.select(select);
-        statement.exec((err, items) => {
+            options.select = select;
+        this._collection.find(filter, options).toArray((err, items) => {
             if (err) {
                 callback(err, null);
                 return;
             }
             if (items != null)
-                this._logger.trace(correlationId, "Retrieved %d from %s", items.length, this._collection);
+                this._logger.trace(correlationId, "Retrieved %d from %s", items.length, this._collectionName);
             items = _.map(items, this.convertToPublic);
             if (pagingEnabled) {
-                this._model.countDocuments(filter, (err, count) => {
+                this._collection.countDocuments(filter, (err, count) => {
                     if (err) {
                         callback(err, null);
                         return;
@@ -196,19 +200,19 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
      * @param callback         callback function that receives a data list or error.
      */
     getListByFilter(correlationId, filter, sort, select, callback) {
-        // Configure statement
-        let statement = this._model.find(filter);
+        // Configure options
+        let options = {};
         if (sort && !_.isEmpty(sort))
-            statement.sort(sort);
+            options.sort = sort;
         if (select && !_.isEmpty(select))
-            statement.select(select);
-        statement.exec((err, items) => {
+            options.select = select;
+        this._collection.find(filter, options).toArray((err, items) => {
             if (err) {
                 callback(err, null);
                 return;
             }
             if (items != null)
-                this._logger.trace(correlationId, "Retrieved %d from %s", items.length, this._collection);
+                this._logger.trace(correlationId, "Retrieved %d from %s", items.length, this._collectionName);
             items = _.map(items, this.convertToPublic);
             callback(null, items);
         });
@@ -234,9 +238,10 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
      * @param callback          callback function that receives data item or error.
      */
     getOneById(correlationId, id, callback) {
-        this._model.findById(id, (err, item) => {
+        let filter = { _id: id };
+        this._collection.findOne(filter, (err, item) => {
             if (!err)
-                this._logger.trace(correlationId, "Retrieved from %s by id = %s", this._collection, id);
+                this._logger.trace(correlationId, "Retrieved from %s by id = %s", this._collectionName, id);
             item = this.convertToPublic(item);
             callback(err, item);
         });
@@ -252,16 +257,17 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
      * @param callback          callback function that receives a random item or error.
      */
     getOneRandom(correlationId, filter, callback) {
-        this._model.countDocuments(filter, (err, count) => {
+        this._collection.countDocuments(filter, (err, count) => {
             if (err) {
                 callback(err, null);
                 return;
             }
             let pos = _.random(0, count - 1);
-            this._model.find(filter)
-                .skip(pos >= 0 ? pos : 0)
-                .limit(1)
-                .exec((err, items) => {
+            let options = {
+                skip: pos >= 0 ? pos : 0,
+                limit: 1,
+            };
+            this._collection.find(filter, options).toArray((err, items) => {
                 let item = (items != null && items.length > 0) ? items[0] : null;
                 item = this.convertToPublic(item);
                 callback(err, item);
@@ -284,10 +290,10 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
         let newItem = _.omit(item, 'id');
         newItem._id = item.id || pip_services3_commons_node_3.IdGenerator.nextLong();
         newItem = this.convertFromPublic(newItem);
-        this._model.create(newItem, (err, newItem) => {
+        this._collection.insertOne(newItem, (err, result) => {
             if (!err)
                 this._logger.trace(correlationId, "Created in %s with id = %s", this._collection, newItem._id);
-            newItem = this.convertToPublic(newItem);
+            newItem = result && result.ops ? this.convertToPublic(result.ops[0]) : null;
             callback(err, newItem);
         });
     }
@@ -313,14 +319,14 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
             _id: newItem._id
         };
         let options = {
-            new: true,
+            returnOriginal: false,
             upsert: true
         };
-        this._model.findOneAndUpdate(filter, newItem, options, (err, newItem) => {
+        this._collection.findOneAndUpdate(filter, newItem, options, (err, result) => {
             if (!err)
                 this._logger.trace(correlationId, "Set in %s with id = %s", this._collection, item.id);
             if (callback) {
-                newItem = this.convertToPublic(newItem);
+                newItem = result ? this.convertToPublic(result.value) : null;
                 callback(err, newItem);
             }
         });
@@ -340,14 +346,16 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
         }
         let newItem = _.omit(item, 'id');
         newItem = this.convertFromPublic(newItem);
+        let filter = { _id: item.id };
+        let update = { $set: newItem };
         let options = {
-            new: true
+            returnOriginal: false
         };
-        this._model.findByIdAndUpdate(item.id, newItem, options, (err, newItem) => {
+        this._collection.findOneAndUpdate(filter, update, options, (err, result) => {
             if (!err)
                 this._logger.trace(correlationId, "Updated in %s with id = %s", this._collection, item.id);
             if (callback) {
-                newItem = this.convertToPublic(newItem);
+                newItem = result ? this.convertToPublic(result.value) : null;
                 callback(err, newItem);
             }
         });
@@ -368,17 +376,16 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
         }
         let newItem = data.getAsObject();
         newItem = this.convertFromPublicPartial(newItem);
-        let setItem = {
-            $set: newItem
-        };
+        let filter = { _id: id };
+        let update = { $set: newItem };
         let options = {
-            new: true
+            returnOriginal: false
         };
-        this._model.findByIdAndUpdate(id, setItem, options, (err, newItem) => {
+        this._collection.findOneAndUpdate(filter, update, options, (err, result) => {
             if (!err)
                 this._logger.trace(correlationId, "Updated partially in %s with id = %s", this._collection, id);
             if (callback) {
-                newItem = this.convertToPublic(newItem);
+                newItem = result ? this.convertToPublic(result.value) : null;
                 callback(err, newItem);
             }
         });
@@ -391,11 +398,12 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
      * @param callback          (optional) callback function that receives deleted item or error.
      */
     deleteById(correlationId, id, callback) {
-        this._model.findByIdAndRemove(id, (err, oldItem) => {
+        let filter = { _id: id };
+        this._collection.findOneAndDelete(filter, (err, result) => {
             if (!err)
                 this._logger.trace(correlationId, "Deleted from %s with id = %s", this._collection, id);
             if (callback) {
-                oldItem = this.convertToPublic(oldItem);
+                let oldItem = result ? this.convertToPublic(result.value) : null;
                 callback(err, oldItem);
             }
         });
@@ -411,7 +419,8 @@ class IdentifiableMongoDbPersistence extends MongoDbPersistence_1.MongoDbPersist
      * @param callback          (optional) callback function that receives error or null for success.
      */
     deleteByFilter(correlationId, filter, callback) {
-        this._model.deleteMany(filter, (err, count) => {
+        this._collection.deleteMany(filter, (err, result) => {
+            let count = result ? result.deletedCount : 0;
             if (!err)
                 this._logger.trace(correlationId, "Deleted %d items from %s", count, this._collection);
             if (callback)